requirements.txt
discord
python-dotenv
pymongo